pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "movement-detection"
version = "0.1.0"
description = "YOLO-based moving vs stopped person detection pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "ultralytics>=8.2.0",
  "supervision>=0.21.0",
  "opencv-python>=4.8.0",
  "numpy>=1.24.0",
  "PyYAML>=6.0.0",
  "fire>=0.6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "black>=24.0.0",
]

[project.scripts]
movement-detection = "movement_detection.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

